+++ apps/ungdomsytelse-deltaker/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @navikt/ungdomsytelse-søknad
 
+## 1.4.7
+
+### Patch Changes
+
+- Erstatte LinkPanel med LinkCard
+
+## 1.4.6
+
+### Patch Changes
+
+- Updated dependencies [ea18e6b]
+    - @navikt/sif-common-soknad-ds@26.1.0
+
 ## 1.4.5
 
 ### Patch Changes